+++ src/collective/fullcalendar/browser/fullcalendar.py
@@ -42,11 +42,11 @@ class IIFullcalendarSettings(Interface):
         required=True,
         vocabulary=SimpleVocabulary.fromValues(
             [
+                "multiMonthYear",
                 "dayGridMonth",
                 "timeGridWeek",
                 "listWeek",
                 "dayGridWeek",
-                "multiMonthYear",
             ]
         ),
         default="dayGridMonth",
@@ -68,7 +68,7 @@ class IIFullcalendarSettings(Interface):
             default=u"Possible values: title, prev, next, prevYear, nextYear, today, dayGridMonth, timeGridWeek, listWeek, dayGridWeek, multiMonthYear",
         ),
         required=False,
-        default="dayGridMonth timeGridWeek listWeek multiMonthYear",
+        default="multiMonthYear dayGridMonth timeGridWeek listWeek",
     )
     weekends = schema.Bool(
         title=_(u"label_weekends", default=u"Show weekends"),
@@ -155,7 +155,9 @@ class IIFullcalendarSettings(Interface):
     # Height of Calendar
     calendarHeight = schema.Int(
         title=_(u"label_calendarHeight", default=u"Calendar height"),
-        description=_(u"help_calendarHeight", default=u"Calendar height in pixels"),
+        description=_(
+            u"help_calendarHeight",
+            default=u"Calendar height in pixels. If left empty, the calendar will adjust its height based on the content."),
         required=False,
     )
     # Enable editing on events